+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@ These are the release notes for the [MediaWiki][mediawiki] [GraphViz extension][
 
 ## GraphViz 1.5.x ## (not released yet)
 * Add tag arguments preparse="dynamic" and preparse="static".
+* Add categories for pages created by this extension.
 
 ## GraphViz 1.4.1 ## (2014-10-21)
 * Fix for [bug 72325](https://bugzilla.wikimedia.org/show_bug.cgi?id=72325).